Refactor RSSEntry into its own file
